Add last changes, since changing to https://api.weatherapi.com/v1, temprature is not ok, but icons are showing

+++ SmartHome/Models/ThermostatDisplay.cs
@@ -4,13 +4,13 @@ namespace SmartHome.Models
 {
     public class ThermostatDisplay : ITemperatureObserver
     {
-        private Action<int> _onTemperatureUpdate;
+        private Action<double> _onTemperatureUpdate;
 
-        public ThermostatDisplay(Action<int> onTemperatureUpdate)
+        public ThermostatDisplay(Action<double> onTemperatureUpdate)
         {
             _onTemperatureUpdate = onTemperatureUpdate;
         }
-        public void UpdateTemperature(int newTemp)
+        public void UpdateTemperature(double newTemp)
         {
             Console.WriteLine($"Thermostat Display: Temperature updated to {newTemp} degrees");
             _onTemperatureUpdate?.Invoke(newTemp);
